requirements.txt
fastapi
uvicorn
google-cloud-bigquery
sentence-transformers
vertexai
google-cloud-logging
python-json-logger
jinja2